Fixed incorrect casting
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@100790 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/align/alngraphic.cpp
@@ -662,8 +662,8 @@ void CAlnGraphic::x_DisplayMaster(int master_len, CNCBINode* center, CHTML_table
 
         spacer_length = (int)(pixel_factor*round_number) 
             - int(kDigitWidth*(previous_digitstr.size() 
-                           - previous_digitstr.size()/2) 
-            - kDigitWidth*(digit_str.size()/2));
+                           - previous_digitstr.size()/2)) 
+            - int(kDigitWidth*(digit_str.size()/2));
         previous_digitstr = digit_str;
 
         image = new CHTML_img(m_ImagePath + kGifWhite, spacer_length, m_BarHeight);
@@ -1034,8 +1034,8 @@ string CAlnGraphic::x_FormatScale(void)
 
         int spacer_length = (int)(m_Pixel_factor*m_Round_number) 
             - int(kDigitWidth*(previous_digitstr.size() 
-                           - previous_digitstr.size()/2) 
-            - kDigitWidth*(digit_str.size()/2));
+                           - previous_digitstr.size()/2)) 
+            - int(kDigitWidth*(digit_str.size()/2));
         previous_digitstr = digit_str;
 
         grScaleDigits += x_FormatScaleDigit(digit_str,spacer_length);
